requirements.txt
django==2.2.9
django-filter==2.3.0
djongo==1.2.36
sqlparse==0.2.4
Pillow==7.1.2
dnspython==1.16.0
django-multiselectfield==0.1.12
django-user-sessions==1.7.1
django-jquery==3.1.0
django-simple-history==2.11.0
xhtml2pdf==0.2.4
reportlab==3.5.46
six==1.15.0
html5lib==1.1
PyPDF2==1.26.0
urllib3==1.25.9
webencodings==0.5.1
selenium==3.141.0